+++ routes/web.php
@@ -99,6 +99,9 @@ Route::post('/cart/update_cart', 				'CartController@updateCart');//pending
 Route::post('/cart/{id}/increase', 				'CartController@increase');
 Route::post('/cart/{id}/decrease',				'CartController@decrease');
 
+Route::get('/cart/{id}/component/{cc_id}',      'CartComponentController@show');
+
+
 
 //CART points
 //show cart